revert: undo .js → .jsx renames in commons-ui-core, commons-ui-next, and commons-ui-payload
These renames were introduced under an earlier approach to work around a
tsx parse error. The final fix (subpath exports in @commons-ui/payload)
never loads the React component chain during generate:types, making the
renames unnecessary. Reverting keeps the PR diff focused on the actual fix.
