+++ packages/commons-ui-core/src/Copyright/Copyright.js
