+++ packages/commons-ui-core/src/RichTypography/RichTypography.js
