+++ packages/commons-ui-next/src/NextImageButton/index.js
@@ -1,1 +1,1 @@
-export { default } from "./NextImageButton.jsx";
+export { default } from "./NextImageButton";
