+++ packages/commons-ui-payload/src/LexicalRichText/converter.js
