+++ packages/commons-ui-payload/src/RichText/index.js
@@ -1,3 +1,3 @@
-import RichText from "./RichText.jsx";
+import RichText from "./RichText";
 
 export default RichText;
